second-commit

+++ src/main/java/com/arun/Employee_crud/controller/EmployeeController.java
@@ -3,6 +3,7 @@ package com.arun.Employee_crud.controller;
 import com.arun.Employee_crud.model.Employee;
 import com.arun.Employee_crud.service.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,53 +16,68 @@ public class EmployeeController {
     private  EmployeeService employeeService;
 
     @GetMapping("/employee")
-    public List<Employee> findall() {
-        return employeeService.findAll();
+    public ResponseEntity<List<Employee> >findall() {
+        return ResponseEntity.ok(employeeService.findAll());
     }
 
     @GetMapping("/employee/{id}")
-    public Optional<Employee> findOne(@PathVariable int id) {
-        return employeeService.findOne(id);
+    public ResponseEntity<Optional<Employee> >findOne(@PathVariable int id) {
+
+        if (employeeService.findOne(id).isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(employeeService.findOne(id));
     }
 
     @PostMapping("/employee")
-    public String Create(@RequestBody Employee employee) {
-        if (employee == null) {
-            return "Employee Creation Failed!";
+    public ResponseEntity<String> Create(@RequestBody Employee employee) {
+        if ( employee.getEmail().isEmpty() || employee.getName().isEmpty()
+                || employee.getPhone().isEmpty())
+        {
+                return ResponseEntity.badRequest().body("Name, Phone and Email are required!");
+        } else if (employeeService.existsByEmail(employee.getEmail())) {
+                return ResponseEntity.badRequest().body("Email already exists!");
         }
         employeeService.save(employee);
-        return "Employee Created Successfully!";
+                return ResponseEntity.ok("Employee Created Successfully!");
     }
 
 
 
 
     @GetMapping("/employee/{id}/delete")
-    public String delete(@PathVariable int id) {
+    public ResponseEntity<String> delete(@PathVariable int id) {
+
+        if (employeeService.findOne(id).isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+
         employeeService.delete(id);
-       return "Employee Deleted Successfully!";
+       return ResponseEntity.ok("Employee Deleted Successfully!");
     }
 
     @GetMapping("/employee/search")
-    public List<Employee> search(@RequestParam("s") String s) {
-        if (s.equals("")) {
-            return employeeService.findAll();
+    public ResponseEntity<List<Employee>> search(@RequestParam("s") String s) {
+        if (s.isEmpty()) {
+              return ResponseEntity.badRequest().build();
         }
-
-        return employeeService.search(s);
+        else if (employeeService.search(s).isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(employeeService.search(s));
 
     }
 
     @PutMapping("/employee/{id}")
-    public String update(@PathVariable int id, @RequestBody Employee employee) {
+    public ResponseEntity<String> update(@PathVariable int id, @RequestBody Employee employee) {
         Optional<Employee> emp = employeeService.findOne(id);
         if (emp.isPresent()) {
             Employee e = emp.get();
             e.setName(employee.getName());
             e.setPhone(employee.getPhone());
             employeeService.save(e);
-            return "Employee Updated Successfully!";
+            return ResponseEntity.ok("Employee Updated Successfully!");
         }
-        return "Employee Not Found!";
+        return ResponseEntity.notFound().build();
     }
 }
